requirements.txt
Django==5.1.3
python-decouple==3.8
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==5.0.0
factory-boy==3.3.1
